+++ server/core_storage.go
@@ -574,7 +574,8 @@ func storageWriteObjects(ctx context.Context, logger *zap.Logger, metrics Metric
 		var resultVersion string
 		var createTime time.Time
 		var updateTime time.Time
-		err := br.QueryRow().Scan(&resultRead, &resultWrite, &resultVersion, &createTime, &updateTime)
+		var isUpsert bool
+		err := br.QueryRow().Scan(&resultRead, &resultWrite, &resultVersion, &createTime, &updateTime, &isUpsert)
 		var pgErr *pgconn.PgError
 		if err != nil && errors.As(err, &pgErr) {
 			if pgErr.Code == dbErrorUniqueViolation {
@@ -584,29 +585,27 @@ func storageWriteObjects(ctx context.Context, logger *zap.Logger, metrics Metric
 			return nil, err
 		} else if err == pgx.ErrNoRows {
 			// Not every case from storagePrepWriteObject can return NoRows, but those
-			// which do it is always ErrStorageRejectedVersion
+			// which do are always ErrStorageRejectedVersion
 			metrics.StorageWriteRejectCount(map[string]string{"collection": object.Collection, "reason": "version"}, 1)
 			return nil, runtime.ErrStorageRejectedVersion
 		} else if err != nil {
 			return nil, err
 		}
 
-		if !(op.permissionRead() == resultRead &&
-			op.permissionWrite() == resultWrite &&
-			op.expectedVersion() == resultVersion) {
-			// Write failed, it can happen for 3 reasons:
-			// - constraint violation on insert (handles elsewhere)
-			// - permission: non authoritative write & original row write != 1
-			// - version mismatch
+		if !isUpsert {
+			// Conditions for successful update or insert were not met in the query, following checks are needed to disambiguate
+			// which error to return.
 			if !authoritativeWrite && resultWrite != 1 {
+				// - permission: non-authoritative write & original row write != 1
 				metrics.StorageWriteRejectCount(map[string]string{"collection": object.Collection, "reason": "permission"}, 1)
 				return nil, runtime.ErrStorageRejectedPermission
 			} else {
-				// version check failed
+				// - version mismatch
 				metrics.StorageWriteRejectCount(map[string]string{"collection": object.Collection, "reason": "version"}, 1)
 				return nil, runtime.ErrStorageRejectedVersion
 			}
 		}
+
 		ack := &api.StorageObjectAck{
 			Collection: object.Collection,
 			Key:        object.Key,
@@ -652,19 +651,18 @@ func storagePrepBatch(batch *pgx.Batch, authoritativeWrite bool, op *StorageOpWr
 			UPDATE storage SET value = $4, version = $5, read = $6, write = $7, update_time = now()
 			WHERE collection = $1 AND key = $2 AND user_id = $3 AND version = $8
 		` + writeCheck + `
-			AND NOT (storage.version = $5 AND storage.read = $6 AND storage.write = $7) -- micro optimization: don't update row unnecessary
 			RETURNING read, write, version, create_time, update_time
 		)
-		(SELECT read, write, version, create_time, update_time from upd)
+		(SELECT read, write, version, create_time, update_time, true AS update FROM upd)
 		UNION ALL
-		(SELECT read, write, version, create_time, update_time FROM storage WHERE collection = $1 and key = $2 and user_id = $3)
+		(SELECT read, write, version, create_time, update_time, false AS update FROM storage WHERE collection = $1 and key = $2 and user_id = $3 AND NOT EXISTS (SELECT 1 FROM upd))
 		LIMIT 1`
 
 		params = append(params, object.Version)
 
 		// Outcomes:
 		// - No rows: if no rows returned, then object was not found in DB and can't be updated
-		// - We have row returned, but now we need to know if update happened, that is its WHERE matched
+		// - We have row returned, but now we need to know if update happened, that is if WHERE matched
 		//	 * write != 1 means no permission to write
 		//	 * dbVersion != original version means OCC failure
 
@@ -681,16 +679,15 @@ func storagePrepBatch(batch *pgx.Batch, authoritativeWrite bool, op *StorageOpWr
 			ON CONFLICT (collection, key, user_id) DO
 				UPDATE SET value = $4, version = $5, read = $6, write = $7, update_time = now()
 				WHERE TRUE` + writeCheck + `
-				AND NOT (storage.version = $5 AND storage.read = $6 AND storage.write = $7) -- micro optimization: don't update row unnecessary
 			RETURNING read, write, version, create_time, update_time
 		)
-		(SELECT read, write, version, create_time, update_time from upd)
+		(SELECT read, write, version, create_time, update_time, true AS upsert FROM upd)
 		UNION ALL
-		(SELECT read, write, version, create_time, update_time FROM storage WHERE collection = $1 and key = $2 and user_id = $3)
+		(SELECT read, write, version, create_time, update_time, false AS upsert FROM storage WHERE collection = $1 and key = $2 and user_id = $3 AND NOT EXISTS (SELECT 1 FROM upd))
 		LIMIT 1`
 
 		// Outcomes:
-		// - Row is always returned, need to know if update happened, that is its WHERE matches
+		// - Row is always returned, need to know if update happened, that WHERE matches
 		// - write != 1 means no permission to write
 
 	case object.Version == "*":
@@ -699,7 +696,7 @@ func storagePrepBatch(batch *pgx.Batch, authoritativeWrite bool, op *StorageOpWr
 		query = `
 		INSERT INTO storage (collection, key, user_id, value, version, read, write, create_time, update_time)
 		VALUES ($1, $2, $3, $4, $5, $6, $7, now(), now())
-		RETURNING read, write, version, create_time, update_time`
+		RETURNING read, write, version, create_time, update_time, true AS upsert`
 
 		// Outcomes:
 		// - NoRows - insert failed due to constraint violation (concurrent insert)