+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 ## [Unreleased]
 ### Fixed
 - Skip Google refund handling for deleted users.
+- Fix storage engine version check regression.
 
 ## [3.20.0] - 2023-12-15
 ### Changed